+++ apps/api/src/app/order/order.service.ts
@@ -14,6 +14,7 @@ import {
 } from '@ghostfolio/common/config';
 import { getAssetProfileIdentifier } from '@ghostfolio/common/helper';
 import {
+  ActivitiesResponse,
   AssetProfileIdentifier,
   EnhancedSymbolProfile,
   Filter
@@ -36,8 +37,6 @@ import { isUUID } from 'class-validator';
 import { endOfToday, isAfter } from 'date-fns';
 import { groupBy, uniqBy } from 'lodash';
 import { v4 as uuidv4 } from 'uuid';
-
-import { Activities } from './interfaces/activities.interface';
 
 @Injectable()
 export class OrderService {
@@ -345,7 +344,7 @@ export class OrderService {
     userCurrency: string;
     userId: string;
     withExcludedAccountsAndActivities?: boolean;
-  }): Promise<Activities> {
+  }): Promise<ActivitiesResponse> {
     let orderBy: Prisma.Enumerable<Prisma.OrderOrderByWithRelationInput> = [
       { date: 'asc' }
     ];