+++ libs/common/src/lib/interfaces/index.ts
@@ -38,6 +38,11 @@ import type { Product } from './product';
 import type { AccessTokenResponse } from './responses/access-token-response.interface';
 import type { AccountBalancesResponse } from './responses/account-balances-response.interface';
 import type { AccountsResponse } from './responses/accounts-response.interface';
+import type {
+  ActivitiesResponse,
+  ActivityResponse,
+  ActivityResponseError
+} from './responses/activity-response.interface';
 import type { AiPromptResponse } from './responses/ai-prompt-response.interface';
 import type { ApiKeyResponse } from './responses/api-key-response.interface';
 import type { BenchmarkMarketDataDetailsResponse } from './responses/benchmark-market-data-details-response.interface';
@@ -80,6 +85,9 @@ export {
   AccountBalance,
   AccountBalancesResponse,
   AccountsResponse,
+  ActivitiesResponse,
+  ActivityResponse,
+  ActivityResponseError,
   AdminData,
   AdminJobs,
   AdminMarketData,